requirements.txt
click==8.1.7
drawsvg==2.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2==2.9.9
Pygments==2.17.2
PyYAML==6.0.1
rich==13.7.0
typer==0.9.0
typing_extensions==4.9.0
